refactor(settings_view): split theme preview helpers into theme_preview.rs
Lifts the theme rendering helpers out of mod.rs into a new
`settings_view::theme_preview` child module:

  * `pub(super)` entry points:
    - `theme_swatches(&Theme)` — compact 8-color strip rendered
      at the left of each installed-themes row.
    - `theme_preview_block(&Theme)` — full-fidelity preview
      dialog body, sample lines + paired fg/bg blocks.
  * Module-private helpers:
    - `preview_sample_lines()` — the canned sample-output rows
      the preview block renders against each theme's palette.
    - `parse_theme_color(&str)` — `#RRGGBB` / `#RGB` / `#RRGGBBAA`
      → `0xRRGGBBAA` u32 used by gpui's `rgba`.

About 235 lines of self-contained colour-rendering code that
has no callers outside `settings_view`. The `SharedString`
import comes along on theme_preview.rs (used inline by
`theme_preview_block`); the `gpui::SharedString` line stays in
mod.rs because settings_view's keystroke-capture path still
needs it.

mod.rs drops by ~235 lines (3,452 → 3,215).

Verified: `cargo clippy --all-targets -- -D warnings` and
`cargo test` both pass (63 tests, unchanged).
